+++ src/Renderer/Models/LineModel.h
@@ -30,9 +30,9 @@ public:
 
     void SetRenderer(Renderer *renderer) override;
 
-    void SetModelId(unsigned long long int id) override;
+    void SetModelId(long long int id) override;
 
-    size_t & GetModelId() override;
+    long long int & GetModelId() override;
 
     void ReceiveCommand(std::unique_ptr<RenderMessage> message) override;
